+++ test/test_helper.rb
@@ -11,6 +11,8 @@ Minitest::Test = Minitest::Unit::TestCase unless defined?(Minitest::Test)
 File.delete("elasticsearch.log") if File.exist?("elasticsearch.log")
 Searchkick.client.transport.logger = Logger.new("elasticsearch.log")
 
+puts "Running against Elasticsearch #{Searchkick.server_version}"
+
 I18n.config.enforce_available_locales = true
 
 ActiveJob::Base.logger = nil if defined?(ActiveJob)
@@ -49,6 +51,7 @@ if defined?(Mongoid)
     field :in_stock, type: Boolean
     field :backordered, type: Boolean
     field :orders_count, type: Integer
+    field :found_rate, type: BigDecimal
     field :price, type: Integer
     field :color
     field :latitude, type: BigDecimal
@@ -83,27 +86,32 @@ elsif defined?(NoBrainer)
     include NoBrainer::Document
     include NoBrainer::Document::Timestamps
 
+    field :id,           type: Object
     field :name,         type: String
-    field :store_id,     type: Integer
     field :in_stock,     type: Boolean
     field :backordered,  type: Boolean
     field :orders_count, type: Integer
+    field :found_rate
     field :price,        type: Integer
     field :color,        type: String
     field :latitude
     field :longitude
     field :description,  type: String
+
+    belongs_to :store, validates: false
   end
 
   class Store
     include NoBrainer::Document
 
+    field :id,   type: Object
     field :name, type: String
   end
 
   class Animal
     include NoBrainer::Document
 
+    field :id,   type: Object
     field :name, type: String
   end
 
@@ -133,6 +141,7 @@ else
     t.boolean :in_stock
     t.boolean :backordered
     t.integer :orders_count
+    t.decimal :found_rate
     t.integer :price
     t.string :color
     t.decimal :latitude, precision: 10, scale: 7
@@ -210,12 +219,15 @@ class Product
 end
 
 class Store
-  searchkick mappings: {
-    store: {
-      properties: {
-        name: {type: "string", analyzer: "keyword"}
+  searchkick \
+    routing: :name,
+    merge_mappings: true,
+    mappings: {
+      store: {
+        properties: {
+          name: {type: "string", analyzer: "keyword"},
+        }
       }
-    }
   }
 end